Move a variety of packages from dependencies to devDependencies.
Inclusion of @types/node in particular is causing me some headaches due to a
version conflict in a consuming repository. I could fix this with a resolutions
configuration, but I perceive that @types/node should not be a dependency of any
other package in the first place.

A small handful of additional changes have been made:

 - @types/jest has been bumped from 25.2.3 to 27.4.0. This is inline with the
   version of jest in use, which is 27.0.1.

 - @types/mkdirp has been bumped from 1.0.1 to 1.0.2. This is the closest match
   to the version of mkdirp in use, which is 1.0.4.

 - @types/sinon has been bumped from 7.5.2 to 9.0.11. This is the closest match
   to the version of sinon in use, which is 9.2.4.